Bug fixed: 数値フィールドのフィルタが小数点以下に対応できない問題を解消。

+++ Implem.Pleasanter/Libraries/Settings/View.cs
@@ -433,7 +433,7 @@ namespace Implem.Pleasanter.Libraries.Settings
                     name: columnName,
                     _operator: " in ({0})".Params(param
                         .Where(o => o != "\t")
-                        .Select(o => o.ToLong())
+                        .Select(o => o.ToDecimal())
                         .Join()))
                 : null;
         }
